remove theme provider use global i18n function instead. remove extra console.log. improve createI18n utils
