+++ .eslintrc
@@ -25,6 +25,9 @@
         "config": "webpack_configs/devServer.config.js"
       }
     }
+  },
+  "globals": {
+    "i18n": false
   }
 }
