+++ app/containers/Root/index.jsx
@@ -5,7 +5,6 @@ import { Provider, connect } from 'react-redux'
 import store from '../../store'
 import IDE from '../IDE.jsx'
 import WorkspaceList from '../../components/Workspace'
-import ThemeProvider from '../../components/ThemeProvider'
 import { initState } from './actions'
 
 
@@ -15,13 +14,10 @@ const codeTranslate = (language) => {
     Chinese: 'zh_CN'
   }
   if (Array.isArray(language)) {
-    console.log('here', language)
     const a = language.find(lan => {
       if (lan.includes('-')) lan = lan.replace(/-/g, '_')
-      console.log('hrere', Object.keys(dic).map(e => dic[e]).includes(lan), lan)
       return Object.keys(dic).map(e => dic[e]).includes(lan)
     })
-    console.log('12312313', a.replace(/-/g, '_'))
     return a.replace(/-/g, '_')
   }
   if (dic[language]) return dic[language]
@@ -59,9 +55,7 @@ Root = connect(
 export default () => {
   return (
     <Provider store={store}>
-      <ThemeProvider>
-        <Root id='root-container' />
-      </ThemeProvider>
+      <Root id='root-container' />
     </Provider>
   )
 }
